Se arreglan caleeeta de weas como tambien se agregan varios metodos para cada cosa

+++ src/dijkstra/AbstractDijkstra.java
@@ -3,11 +3,13 @@ package dijkstra;
 import data.Graph;
 
 public abstract class AbstractDijkstra {
+
     private Graph graph;
 
-    public AbstractDijkstra(Graph graph) {
+    public AbstractDijkstra(Graph graph)
+    {
         this.graph = graph;
     }
 
-    public abstract float shortestPath();
+    public abstract double shortestPath();
 }
